Handling errors

+++ src/app/page.js
@@ -74,22 +74,28 @@ export default function Home() {
   };
 
   const handleAddToCart = (product) => {
-    // Get the current cart from localStorage or initialize an empty array if not present
     let cart = JSON.parse(localStorage.getItem('cart')) || [];
 
-    // Add the full product object to the cart array
-    cart.push(product);
+    const existingProductIndex = cart.findIndex(item => item.id === product.id);
+    if (existingProductIndex !== -1) {
+      toast.error(
+        "Product already added to cart 🛒",
+        {
+          duration: 3000
+        }
+      );
+    } else {
 
-    // Save the updated cart array back to localStorage
-    localStorage.setItem('cart', JSON.stringify(cart));
+      cart.push(product);
+      localStorage.setItem('cart', JSON.stringify(cart));
 
-    // Notify the user
-    toast.success(
-      "Product added to cart 🛒",
-      {
-        duration: 4000
-      }
-    );
+      toast.success(
+        "Product added to cart 🛒",
+        {
+          duration: 3000
+        }
+      );
+    }
   };
